+++ packages/orq-rc/RELEASES.md
@@ -7199,3 +7199,13 @@ Based on:
 - [typescript v4.4.0-rc.37] packages/orq-rc
 ### Releases
 - [NPM v4.4.0-rc.37] https://www.npmjs.com/package/@orq-ai/node/v/4.4.0-rc.37 - packages/orq-rc
+
+## 2026-02-24 09:26:55
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.725.0 (2.836.5) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.5.0-rc.0] packages/orq-rc
+### Releases
+- [NPM v4.5.0-rc.0] https://www.npmjs.com/package/@orq-ai/node/v/4.5.0-rc.0 - packages/orq-rc